Revert "Bug Fix: StartingVersionForPhaseId contains incorrect values (#101)"
This reverts commit 227f1b997e393de0783baa9e37377ac871e9df38.

+++ packages/perennial-oracle/contracts/types/ChainlinkRound.sol
@@ -7,7 +7,7 @@ import "@chainlink/contracts/src/v0.8/interfaces/AggregatorV3Interface.sol";
 struct ChainlinkRound {
     uint256 timestamp;
     int256 answer;
-    uint256 roundId;
+    uint80 roundId;
 }
 using ChainlinkRoundLib for ChainlinkRound global;
